requirements.txt
GitPython==3.1.43
PyGithub==2.4.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
python-gitlab>=3.0.0
